feat(ci): add prepareAionuiBackend integration for CI builds
- Add prepareAionuiBackend() call in build-with-builder.js before prepareBundledBun
- Add AIONUI_BACKEND_ALLOW_MISSING env var for transition control
- Extract prepareAionuiBackend as reusable module in shared-scripts/src/prepare-aionui-backend.js
- Add prepareAionuiBackend step in CI workflow (_build-reusable.yml)
- Support early exit with skip manifest when allowMissing=true and release unavailable
- Hard fail when allowMissing=false (default for CI after backend release ready)

M7 Phase 1-4: build integration + CI workflow + ALLOW_MISSING transition switch
