+++ .github/workflows/_build-reusable.yml
@@ -304,6 +304,15 @@ jobs:
           MSVS_VERSION: 2022
           GYP_MSVS_VERSION: 2022
 
+      - name: Prepare aionui-backend binary
+        shell: bash
+        run: node scripts/prepareAionuiBackend.js
+        env:
+          AIONUI_BACKEND_VERSION: latest
+          AIONUI_BACKEND_ALLOW_MISSING: '1'
+          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+
       - name: Build with electron-builder (Windows)
         if: startsWith(matrix.platform, 'windows')
         id: windows-build
