+++ packages/shared-scripts/src/prepare-aionui-backend.js
@@ -1,0 +1,326 @@
+/**
+ * Prepare aionui-backend binary for packaging.
+ *
+ * Resolution order:
+ *  1. GitHub release download (requires version or defaults to "latest")
+ *
+ * Output: {projectRoot}/resources/bundled-aionui-backend/{platform}-{arch}/aionui-backend[.exe]
+ *
+ * @module prepare-aionui-backend
+ */
+
+const { execSync, execFileSync } = require('child_process');
+const fs = require('fs');
+const os = require('os');
+const path = require('path');
+
+const GITHUB_OWNER = 'iOfficeAI';
+const GITHUB_REPO = 'aionui-backend';
+
+// ---------------------------------------------------------------------------
+// Helpers
+// ---------------------------------------------------------------------------
+
+function ensureDirectory(dirPath) {
+  if (!fs.existsSync(dirPath)) {
+    fs.mkdirSync(dirPath, { recursive: true });
+  }
+}
+
+function removeDirectorySafe(dirPath) {
+  fs.rmSync(dirPath, { recursive: true, force: true });
+}
+
+function copyFileSafe(sourcePath, targetPath) {
+  ensureDirectory(path.dirname(targetPath));
+  fs.copyFileSync(sourcePath, targetPath);
+}
+
+function ensureExecutableMode(filePath) {
+  if (process.platform === 'win32') return;
+  try {
+    fs.chmodSync(filePath, 0o755);
+  } catch {}
+}
+
+function writeJson(filePath, payload) {
+  fs.writeFileSync(filePath, JSON.stringify(payload, null, 2) + '\n', 'utf-8');
+}
+
+function getBinaryName(platform) {
+  return platform === 'win32' ? 'aionui-backend.exe' : 'aionui-backend';
+}
+
+// ---------------------------------------------------------------------------
+// Source resolvers
+// ---------------------------------------------------------------------------
+
+/**
+ * Resolve the actual version tag when "latest" is requested.
+ * Uses GitHub API via `gh` CLI (needs GH_TOKEN in CI) or falls back to
+ * `curl` with an optional Authorization header (GITHUB_TOKEN / GH_TOKEN).
+ */
+function resolveLatestTag() {
+  const token = process.env.GH_TOKEN || process.env.GITHUB_TOKEN || '';
+
+  // 1. Try gh CLI (honours GH_TOKEN automatically)
+  try {
+    const out = execSync(`gh api repos/${GITHUB_OWNER}/${GITHUB_REPO}/releases/latest --jq .tag_name`, {
+      encoding: 'utf-8',
+      timeout: 15000,
+    }).trim();
+    if (out) return out;
+  } catch {
+    // gh CLI not available or no token — fall back to curl
+  }
+
+  // 2. Curl with optional token to avoid rate-limit 403
+  try {
+    const authArgs = token ? ['-H', `Authorization: token ${token}`] : [];
+    const args = ['-fsSL', ...authArgs, `https://api.github.com/repos/${GITHUB_OWNER}/${GITHUB_REPO}/releases/latest`];
+    const out = execFileSync('curl', args, { encoding: 'utf-8', timeout: 15000 });
+    const tag = JSON.parse(out).tag_name;
+    if (tag) return tag;
+  } catch {
+    // network issue or rate-limited
+  }
+
+  return null;
+}
+
+/**
+ * Build the release asset filename for the given platform/arch/tag.
+ *
+ * Expected asset naming convention:
+ *   aionui-backend-v0.1.0-aarch64-apple-darwin.tar.gz
+ */
+function getAssetName(platform, arch, tag) {
+  const archMap = { x64: 'x86_64', arm64: 'aarch64' };
+  const platformMap = {
+    darwin: 'apple-darwin',
+    linux: 'unknown-linux-gnu',
+    win32: 'pc-windows-msvc',
+  };
+  const normalizedArch = archMap[arch];
+  const normalizedPlatform = platformMap[platform];
+  if (!normalizedArch || !normalizedPlatform) return null;
+  const ext = platform === 'win32' ? '.zip' : '.tar.gz';
+  return `aionui-backend-${tag}-${normalizedArch}-${normalizedPlatform}${ext}`;
+}
+
+function getDownloadUrl(assetName, tag) {
+  return `https://github.com/${GITHUB_OWNER}/${GITHUB_REPO}/releases/download/${tag}/${assetName}`;
+}
+
+function downloadFile(url, outputPath) {
+  console.log(`  Downloading aionui-backend from ${url}`);
+  if (process.platform === 'win32') {
+    const ps = `$ProgressPreference='SilentlyContinue'; Invoke-WebRequest -Uri '${url}' -OutFile '${outputPath.replace(/'/g, "''")}'`;
+    execFileSync('powershell', ['-NoProfile', '-NonInteractive', '-Command', ps], {
+      timeout: 120000,
+    });
+    return;
+  }
+  try {
+    execFileSync('curl', ['-L', '--fail', '--silent', '--show-error', '-o', outputPath, url], { timeout: 120000 });
+  } catch {
+    execFileSync('wget', ['-q', '-O', outputPath, url], { timeout: 120000 });
+  }
+}
+
+function extractArchive(archivePath, outputDir, platform) {
+  ensureDirectory(outputDir);
+  if (platform === 'win32' || archivePath.endsWith('.zip')) {
+    if (platform === 'win32') {
+      const ps = `Expand-Archive -LiteralPath '${archivePath.replace(/'/g, "''")}' -DestinationPath '${outputDir.replace(/'/g, "''")}' -Force`;
+      execFileSync('powershell', ['-NoProfile', '-NonInteractive', '-Command', ps]);
+    } else {
+      execFileSync('unzip', ['-o', archivePath, '-d', outputDir]);
+    }
+  } else {
+    execFileSync('tar', ['-xzf', archivePath, '-C', outputDir]);
+  }
+}
+
+function findBinaryInDir(dir, binaryName) {
+  const entries = fs.readdirSync(dir, { withFileTypes: true });
+  for (const entry of entries) {
+    const fullPath = path.join(dir, entry.name);
+    if (entry.isFile() && entry.name === binaryName) return fullPath;
+    if (entry.isDirectory()) {
+      const found = findBinaryInDir(fullPath, binaryName);
+      if (found) return found;
+    }
+  }
+  return null;
+}
+
+function downloadAndExtract(platform, arch, tag) {
+  const assetName = getAssetName(platform, arch, tag);
+  if (!assetName) {
+    throw new Error(`Unsupported aionui-backend target: ${platform}-${arch}`);
+  }
+
+  const url = getDownloadUrl(assetName, tag);
+  const tempDir = path.join(os.tmpdir(), 'aionui-backend-prepare', tag, `${platform}-${arch}`);
+  const archivePath = path.join(tempDir, assetName);
+  const extractDir = path.join(tempDir, 'extracted');
+
+  removeDirectorySafe(tempDir);
+  ensureDirectory(tempDir);
+
+  downloadFile(url, archivePath);
+  extractArchive(archivePath, extractDir, platform);
+
+  const binaryName = getBinaryName(platform);
+  const binaryPath = findBinaryInDir(extractDir, binaryName);
+  if (!binaryPath) {
+    throw new Error(`Binary ${binaryName} not found in downloaded archive`);
+  }
+
+  return { binaryPath, tempDir, url };
+}
+
+// ---------------------------------------------------------------------------
+// Main export
+// ---------------------------------------------------------------------------
+
+/**
+ * Prepare aionui-backend binary for packaging.
+ *
+ * @param {object} options - Configuration options
+ * @param {string} options.projectRoot - Project root directory
+ * @param {string} options.platform - Target platform (process.platform)
+ * @param {string} options.arch - Target architecture (process.arch)
+ * @param {string} options.version - Backend version (default: 'latest')
+ * @param {boolean} options.allowMissing - Allow missing backend (default: false)
+ * @returns {{ prepared: boolean; dir?: string; sourceType?: string; reason?: string }}
+ */
+function prepareAionuiBackend(options) {
+  const { projectRoot, platform, arch, version = 'latest', allowMissing = false } = options;
+  const runtimeKey = `${platform}-${arch}`;
+
+  // Resolve the actual version tag — asset filenames include the tag
+  let tag;
+  if (version === 'latest') {
+    const resolved = resolveLatestTag();
+    if (!resolved) {
+      if (allowMissing) {
+        // Write skip manifest and return early
+        const targetDir = path.join(projectRoot, 'resources', 'bundled-aionui-backend', runtimeKey);
+        const binaryName = getBinaryName(platform);
+        ensureDirectory(targetDir);
+        const manifest = {
+          platform,
+          arch,
+          version: 'unknown',
+          generatedAt: new Date().toISOString(),
+          sourceType: 'none',
+          source: {},
+          files: [],
+          skipped: true,
+          reason: 'Failed to resolve latest aionui-backend release tag from GitHub API',
+        };
+        writeJson(path.join(targetDir, 'manifest.json'), manifest);
+        console.warn(
+          `  aionui-backend not found — skipping bundle (AIONUI_BACKEND_ALLOW_MISSING=1, backend will not be available in packaged app)`
+        );
+        return { prepared: false, reason: 'not_found' };
+      } else {
+        throw new Error('Failed to resolve latest aionui-backend release tag from GitHub API');
+      }
+    }
+    tag = resolved;
+    console.log(`Resolved aionui-backend "latest" → ${tag}`);
+  } else {
+    tag = version.startsWith('v') ? version : `v${version}`;
+  }
+
+  const targetDir = path.join(projectRoot, 'resources', 'bundled-aionui-backend', runtimeKey);
+  const binaryName = getBinaryName(platform);
+  const targetBinaryPath = path.join(targetDir, binaryName);
+
+  console.log(`Preparing aionui-backend for ${runtimeKey} (version: ${tag})`);
+
+  removeDirectorySafe(targetDir);
+  ensureDirectory(targetDir);
+
+  let sourcePath = null;
+  let sourceType = 'none';
+  let sourceDetail = {};
+  let tempDir = null;
+
+  // 1. Download from GitHub releases
+  if (!sourcePath) {
+    try {
+      const result = downloadAndExtract(platform, arch, tag);
+      sourcePath = result.binaryPath;
+      tempDir = result.tempDir;
+      sourceType = 'download';
+      sourceDetail = { url: result.url };
+      console.log(`  Downloaded from GitHub releases`);
+    } catch (error) {
+      console.warn(`  Download failed: ${error.message}`);
+    }
+  }
+
+  // Write result
+  if (sourcePath) {
+    copyFileSafe(sourcePath, targetBinaryPath);
+    ensureExecutableMode(targetBinaryPath);
+
+    // Get version info from binary
+    let binaryVersion = tag;
+    try {
+      binaryVersion = execSync(`"${targetBinaryPath}" --version`, {
+        encoding: 'utf-8',
+        timeout: 5000,
+      }).trim();
+    } catch {}
+
+    const manifest = {
+      platform,
+      arch,
+      version: binaryVersion,
+      generatedAt: new Date().toISOString(),
+      sourceType,
+      source: sourceDetail,
+      files: [binaryName],
+      skipped: false,
+    };
+
+    writeJson(path.join(targetDir, 'manifest.json'), manifest);
+    console.log(
+      `  Bundled aionui-backend prepared: resources/bundled-aionui-backend/${runtimeKey}/${binaryName} [source=${sourceType}]`
+    );
+
+    if (tempDir) removeDirectorySafe(tempDir);
+    return { prepared: true, dir: targetDir, sourceType };
+  }
+
+  // Not found
+  if (allowMissing) {
+    // Write skip manifest (non-fatal)
+    const manifest = {
+      platform,
+      arch,
+      version: tag,
+      generatedAt: new Date().toISOString(),
+      sourceType: 'none',
+      source: {},
+      files: [],
+      skipped: true,
+      reason: 'aionui-backend binary not found (ensure GitHub release exists)',
+    };
+
+    writeJson(path.join(targetDir, 'manifest.json'), manifest);
+    console.warn(
+      `  aionui-backend not found — skipping bundle (AIONUI_BACKEND_ALLOW_MISSING=1, backend will not be available in packaged app)`
+    );
+    return { prepared: false, reason: 'not_found' };
+  } else {
+    throw new Error('aionui-backend binary not found and AIONUI_BACKEND_ALLOW_MISSING is not set');
+  }
+}
+
+module.exports = { prepareAionuiBackend };
